feat(range-search): add simple range search

+++ stilt/core/src/main/java/com/p4ybill/stilt/parser/SpatialMappingFunction.java
@@ -4,7 +4,7 @@ public class SpatialMappingFunction implements MappingFunction {
 
     @Override
     public int map(Object value, int bits) {
-        float f = (Float) value;
+        double f = (Double) value;
         return (int)f;
     }
 }
